server/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
ultralytics>=8.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
